% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.demo1.R
\docType{data}
\name{dat_demo1}
\alias{dat_demo1}
\title{Demo Data 1}
\format{
A \code{data.frame} with 4 columns, which are:
\describe{
\item{V1}{ continuous varible }
\item{V2}{ continuous varible }
\item{V3}{ categorical variable with 4 levels }
\item{V4}{ categorical variable with 4 levels }
}
}
\usage{
dat_demo1
}
\description{
This dataset contains 100 observations and 4 variables: 2 continuous and 2 categorical with 4 levels, generated from a finite mixture model with Gaussian components.
}
\keyword{datasets}
