% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/den.NB.R
\name{est.cat.dens}
\alias{est.cat.dens}
\title{Categorical Data Density Calculation}
\usage{
est.cat.dens(data)
}
\arguments{
\item{data}{a data frame where rows are observations and columns are factorized categorical variables.}
}
\value{
a list of density value and original data.
}
\description{
This is the function to calculate Naive Bayes density for categorical data.
}
