% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/den.knn.R
\name{knn.density}
\alias{knn.density}
\title{K-th Nearest Neighbor Density Calculation}
\usage{
knn.density(dist, k = NULL)
}
\arguments{
\item{dist}{a numeric distance matrix or data frame object.}

\item{k}{k-th nearest neighbors to be checked to estimate the density of each point.}
}
\value{
a vector of density value.
}
\description{
This is the function to calculate KNN density.
}
\details{
If \code{k} is \code{NULL}, k=n^(4/5), where \code{n} is the number of rows in the \code{dist}.
}
