% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/den.NB.R
\name{NB.cal}
\alias{NB.cal}
\title{Naive Bayes Density Calculation for Mixed-type Data}
\usage{
NB.cal(data, idnum, idcat, con.den = "locfit")
}
\arguments{
\item{data}{a data frame to be clustered, mixed type variables are allowed. Categorical variables should be factorized.}

\item{idnum}{a vector indicating the position of continuous variables.}

\item{idcat}{a vector indicating the position of categorical variables.}

\item{con.den}{method to calculate density of continuous variables. Options: \code{"logspline"} or \code{"locfit"}. Default is ""locfit".}
}
\value{
a vector of density value.
}
\description{
This is the function to calculate Naive Bayes density for mixed-type data.
}
